fix(app-shell): more to utils
